requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
transformers==4.37.0
faiss-cpu==1.7.4
numpy==1.26.3
